+++ ios_mysql_client/sources/MySQLWrapper/MySqlController.m
@@ -29,7 +29,7 @@
 }
 
 - (void)connectToServer:(NSString *)serverName withDbName:(NSString *)dbName
-             withSocket:(NSString *)socket withPort:(NSInteger)port
+             withSocket:(NSString *)socket withPort:(unsigned int)port
            withUsername:(NSString *)username withPassword:(NSString *)password
 {
     [db setSocket:socket];
